requirements.txt
cqlengine==0.21.0
Django==1.7
djangorestframework==3.1.3
gunicorn==19.3.0
django-cors-headers==1.0.0
pika==0.9.14
django-cassandra-engine==0.3.2
cassandra-driver==2.5.1
redis==2.10.3
blist==1.3.6
Pillow==2.8.1
requests==2.7.0
neo4jrestclient==2.1.1
